pyproject.toml
[project]
name = "hd4-fetch-ai-uagent"
version = "0.1.0"
description = "AI Agent Backend with UAgents, PostgreSQL and FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "uagents>=0.22.10",
    "pydantic[email]>=2.5.0",
    "psycopg2-binary>=2.9.9",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "uvicorn>=0.24.0",
    "jwt>=1.4.0",
    "httpx>=0.28.1",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
